+++ includes/api/validation-functions.php
@@ -179,14 +179,21 @@ function validate_richtext_field( $value, array $field ): void {
  */
 function validate_number_field( $value, array $field ): void {
 	if ( is_field_repeatable( $field ) ) {
-		// translators: The name and type of the field.
-		$message = \sprintf( \__( '%1$s must be an array of %2$s', 'atlas-content-modeler' ), $field['name'], $field['type'] );
-
-		validate_array( $value, $message );
+		validate_array(
+			$value,
+			// translators: The field name and field type.
+			\sprintf( \__( '%1$s must be an array of %2$s', 'atlas-content-modeler' ), $field['name'], $field['type'] )
+		);
 
 		if ( is_field_required( $field ) ) {
-			validate_not_empty( $value, $message );
+			validate_not_empty(
+				$value,
+				// translators: The field name.
+				\sprintf( \__( '%s cannot be empty', 'atlas-content-modeler' ), $field['name'] )
+			);
 		}
+
+		validate_row_count_within_repeatable_limits( count( $value ), $field );
 	}
 
 	validate_array_of(
@@ -195,16 +202,25 @@ function validate_number_field( $value, array $field ): void {
 			if ( is_field_required( $field ) ) {
 				validate_not_empty(
 					$field_value,
-					// translators: The name of the field.
-					\sprintf( \__( '%s is required', 'atlas-content-modeler' ), $field['name'] )
+					// translators: The field name.
+					\sprintf( \__( '%s cannot be empty', 'atlas-content-modeler' ), $field['name'] )
 				);
 			}
 
 			validate_number(
 				$field_value,
-				// translators: The name and type of the field.
+				// translators: The field name and field type.
 				\sprintf( \__( '%1$s must be a valid %2$s', 'atlas-content-modeler' ), $field['name'], $field['type'] )
 			);
+
+			validate_number_type(
+				$field_value,
+				$field['numberType'],
+				// translators: The field name and field type.
+				\sprintf( \__( '%1$s must be of type %2$s', 'atlas-content-modeler' ), $field['name'], $field['numberType'] )
+			);
+
+			validate_number_min_max_step( $field_value, $field );
 		}
 	);
 }
@@ -433,6 +449,62 @@ function validate_email_field( $value, array $field ): void {
 }
 
 /**
+ * Validate the number type.
+ *
+ * Type-checks numbers and numeric strings.
+ *
+ * Will not throw for strings such as "test" passed as integer types.
+ * Use `validate_number()` to validate a string as numeric.
+ *
+ * The zero values 0, 0.0, -0 and -0.0 and their string equivalents are
+ * considered valid for all types.
+ *
+ * @param mixed  $value The value as a number or string.
+ * @param mixed  $number_type The number type.
+ * @param string $message Optional. The error message.
+ *
+ * @throws Validation_Exception Exception when value is the invalid type.
+ *
+ * @return void
+ */
+function validate_number_type( $value, $number_type, string $message = '' ): void {
+	$message           = $message ?: \__( 'Value must be a valid number', 'atlas-content-modeler' );
+	$is_numeric_string = is_string( $value ) && is_numeric( $value );
+
+	/**
+	 * Accept zero values for integers and decimals. Helps with JSON responses
+	 * that do not encode 0 correctly as 0.00 or vice versa.
+	 */
+	if (
+		$value === 0
+		|| $value === 0.0
+		|| $is_numeric_string && (float) $value === 0.0
+	) {
+		return;
+	}
+
+	if (
+		$number_type === 'integer'
+		&& (
+			is_float( $value )
+			|| $is_numeric_string && str_contains( $value, '.' )
+		)
+	) {
+		throw new Validation_Exception( $message );
+	}
+
+	if (
+		$number_type === 'decimal'
+		&& (
+			! $is_numeric_string && ! is_float( $value )
+			|| $is_numeric_string && ! str_contains( $value, '.' )
+		)
+	) {
+		throw new Validation_Exception( $message );
+	}
+}
+
+/**
  * Validate for valid number.
  *
  * @param mixed  $value The value.
@@ -447,6 +519,42 @@ function validate_number( $value, string $message = '' ): void {
 
 	if ( ! \is_numeric( $value ) ) {
 		throw new Validation_Exception( $message );
+	}
+}
+
+/**
+ * Validates a number value for number field min and max.
+ *
+ * @param string $value String value of the text field.
+ * @param array  $field Array of values from the field.
+ * @return void
+ */
+function validate_number_min_max_step( $value, $field ) {
+	if ( \is_numeric( $field['minValue'] ?? '' ) ) {
+		validate_min(
+			$value,
+			$field['minValue'],
+			// translators: The field name and field minimum value.
+			\sprintf( \__( '%1$s must be at least %2$d', 'atlas-content-modeler' ), $field['name'], $field['minValue'] )
+		);
+	}
+
+	if ( \is_numeric( $field['maxValue'] ?? '' ) ) {
+		validate_max(
+			$value,
+			$field['maxValue'],
+			// translators: The field name and field maximum value.
+			\sprintf( \__( '%1$s cannot be greater than %2$d', 'atlas-content-modeler' ), $field['name'], $field['maxValue'] )
+		);
+	}
+
+	if ( \is_numeric( $field['step'] ?? '' ) ) {
+		validate_step(
+			$value,
+			$field['step'],
+			// translators: The field name and field step value.
+			\sprintf( \__( '%1$s step must be a multiple of %2$d', 'atlas-content-modeler' ), $field['name'], $field['step'] )
+		);
 	}
 }
 
@@ -618,6 +726,25 @@ function validate_max( $value, int $max, string $message = '' ): void {
 	}
 
 	if ( \is_numeric( $value ) && (float) $value > $max ) {
+		throw new Validation_Exception( $message );
+	}
+}
+
+/**
+ * Validate a number against step.
+ *
+ * @param mixed  $value The value.
+ * @param int    $step The step value.
+ * @param string $message The optional error message.
+ *
+ * @throws Validation_Exception Exception when value is invalid.
+ *
+ * @return void
+ */
+function validate_step( $value, $step, string $message = '' ): void {
+	if ( \is_numeric( $value ) &&
+			\is_numeric( $step ) &&
+			! ( $value % $step === 0 ) ) {
 		throw new Validation_Exception( $message );
 	}
 }