+++ includes/publisher/class-publisher-form-editing-experience.php
@@ -476,8 +476,11 @@ final class FormEditingExperience {
 		foreach ( $posted_values as $key => $value ) {
 			$key = sanitize_text_field( $key );
 
-			// Do not save empty values.
-			if ( empty( $value ) ) {
+			// Delete or ignore empty values, but allow 0 values.
+			if ( empty( $value ) && $value !== '0' && $value !== 0 ) {
+				if ( $update ) {
+					delete_post_meta( $post_id, $key );
+				}
 				continue;
 			}