+++ includes/publisher/js/src/components/Email/RepeatingEmail.jsx
@@ -1,4 +1,6 @@
+/** @jsx jsx */
 import React, { useState, useRef } from "react";
+import { jsx, css } from "@emotion/react";
 import Icon from "../../../../../components/icons";
 import { buildWildcardRegex } from "../../../../../shared-assets/js/validation/emailValidation";
 import TrashIcon from "../../../../../components/icons/TrashIcon";
@@ -119,6 +121,16 @@ const RepeatingEmail = ({
 															: "field"
 													} d-flex flex-row repeater-input mt-0 flex-fill d-lg-flex`}
 												>
+													<span
+														className="px-1 me-2"
+														css={css`
+															font-family: "Open Sans",
+																sans-serif;
+															font-weight: bold;
+														`}
+													>
+														{index + 1}
+													</span>
 													<div className="me-lg-1 repeater-input-container flex-fill">
 														<input
 															type={`${field.type}`}