+++ tests/integration/blueprints/test-blueprint-downloads.php
@@ -108,6 +108,35 @@ class TestBlueprintDownloadTestCases extends WP_UnitTestCase {
 	}
 
 	/**
+	 * @covers ::WPE\AtlasContentModeler\Blueprint\Fetch\save_blueprint_to_upload_dir
+	 */
+	public function test_save_blueprint_to_upload_dir_copies_directory_when_given_a_good_local_directory(): void {
+		if ( ! defined( 'FS_METHOD' ) ) {
+			define( 'FS_METHOD', 'direct' ); // Allows direct filesystem copy operations without FTP/SSH passwords. This only takes effect during testing.
+		}
+
+		$blueprint_folder = __DIR__ . '/test-data/blueprint-good';
+		$destination      = save_blueprint_to_upload_dir( $blueprint_folder, 'blueprint-good' );
+		self::assertTrue( is_readable( $destination ) );
+		self::assertTrue( is_dir( $destination ) );
+	}
+
+	/**
+	 * @covers ::WPE\AtlasContentModeler\Blueprint\Fetch\save_blueprint_to_upload_dir
+	 */
+	public function test_save_blueprint_to_upload_dir_gives_error_when_given_a_bad_local_directory(): void {
+		if ( ! defined( 'FS_METHOD' ) ) {
+			define( 'FS_METHOD', 'direct' ); // Allows direct filesystem copy operations without FTP/SSH passwords. This only takes effect during testing.
+		}
+
+		$non_existent_blueprint_folder = __DIR__ . '/test-data/this-blueprint-does-not-exist';
+		$destination                   = save_blueprint_to_upload_dir( $non_existent_blueprint_folder, 'this-blueprint-does-not-exist' );
+		$this->assertWPError( $destination );
+		self::assertSame( 'acm_blueprint_save_error', $destination->get_error_code() );
+		self::assertStringStartsWith( 'Could not read directory at', $destination->get_error_message() );
+	}
+
+	/**
 	 * Returns test blueprint data.
 	 *
 	 * @return string